refactor: Split the memory service into its own module instead
