+++ packages/@n8n/nodes-langchain/nodes/memory/MemoryBufferWindow/ChatMemoryMessageHistory.test.ts
@@ -5,7 +5,6 @@ import { ChatMemoryMessageHistory } from './ChatMemoryMessageHistory';
 
 describe('ChatMemoryMessageHistory', () => {
 	const createMockMemoryService = (): jest.Mocked<IChatMemoryService> => ({
-		getOwnerId: jest.fn(),
 		getMemory: jest.fn(),
 		addHumanMessage: jest.fn(),
 		addAIMessage: jest.fn(),
