+++ packages/cli/src/modules/chat-memory/chat-memory.entity.ts
