+++ packages/core/src/execution-engine/node-execution-context/supply-data-context.ts
@@ -28,7 +28,7 @@ import {
 	getBinaryDataBuffer,
 	getBinaryHelperFunctions,
 } from './utils/binary-helper-functions';
-import { getChatMemoryHelperFunctions } from './utils/chat-hub-helper-functions';
+import { getChatMemoryHelperFunctions } from './utils/chat-memory-helper-functions';
 import { constructExecutionMetaData } from './utils/construct-execution-metadata';
 import { copyInputItems } from './utils/copy-input-items';
 import { getDataTableHelperFunctions } from './utils/data-table-helper-functions';
@@ -94,7 +94,7 @@ export class SupplyDataContext extends BaseExecuteContext implements ISupplyData
 			...getFileSystemHelperFunctions(node),
 			...getBinaryHelperFunctions(additionalData, workflow.id),
 			...getDataTableHelperFunctions(additionalData, workflow, node),
-			...getChatMemoryHelperFunctions(additionalData, workflow, node, mode),
+			...getChatMemoryHelperFunctions(additionalData, workflow, node),
 			...getDeduplicationHelperFunctions(workflow, node),
 			assertBinaryData: (itemIndex, propertyName) =>
 				assertBinaryData(inputData, node, itemIndex, propertyName, 0, workflow.settings.binaryMode),
